+++ projects/rocprofiler-sdk/source/lib/python/rocpd/source/perfetto.cpp
@@ -180,6 +180,7 @@ write_perfetto(
     const tool::generator<types::sample>&            sample_gen,
     const tool::generator<types::kernel_dispatch>&   kernel_dispatch_gen,
     const tool::generator<types::memory_copies>&     memory_copy_gen,
+    const tool::generator<types::graph_launch>&      graph_launch_gen,
     const tool::generator<types::scratch_memory>&    scratch_memory_gen,
     const tool::generator<types::memory_allocation>& memory_allocation_gen,
     const tool::generator<types::counter>&           counter_collection_gen)
@@ -608,9 +609,56 @@ write_perfetto(
                                   "tid",
                                   itr.tid,
                                   "stream_id",
-                                  itr.stream_id);
+                                  itr.stream_id,
+                                  [&](::perfetto::EventContext ctx) {
+                                      if(itr.graph_exec_id != 0)
+                                      {
+                                          rocprofiler::sdk::add_perfetto_annotation(
+                                              ctx, "graph_exec_id", itr.graph_exec_id);
+                                          rocprofiler::sdk::add_perfetto_annotation(
+                                              ctx, "graph_node_id", itr.graph_node_id);
+                                      }
+                                  });
                 TRACE_EVENT_END(
                     sdk::perfetto_category<sdk::category::memory_copy>::name, *_track, itr.end);
+            }
+            tracing_session->FlushBlocking();
+        }
+
+        for(auto ditr : graph_launch_gen)
+        {
+            for(auto itr : graph_launch_gen.get(ditr))
+            {
+                auto& track = thread_tracks.at(itr.tid);
+                auto  name  = std::string{"hipGraphLaunch"};
+                TRACE_EVENT_BEGIN(sdk::perfetto_category<sdk::category::hip_api>::name,
+                                  ::perfetto::DynamicString{name},
+                                  track,
+                                  itr.start,
+                                  ::perfetto::Flow::Global(itr.stack_id ^ uuid_pid),
+                                  "begin_ns",
+                                  itr.start,
+                                  "end_ns",
+                                  itr.end,
+                                  "delta_ns",
+                                  (itr.end - itr.start),
+                                  "tid",
+                                  itr.tid,
+                                  "kind",
+                                  "GRAPH_LAUNCH",
+                                  "operation",
+                                  name,
+                                  "corr_id",
+                                  itr.stack_id,
+                                  "ancestor_id",
+                                  itr.parent_stack_id,
+                                  "graph_exec_id",
+                                  itr.graph_exec_id,
+                                  "kernel_dispatch_count",
+                                  itr.kernel_dispatch_count);
+                TRACE_EVENT_END(sdk::perfetto_category<sdk::category::hip_api>::name,
+                                track,
+                                itr.end);
             }
             tracing_session->FlushBlocking();
         }
@@ -806,6 +854,14 @@ write_perfetto(
                                   "stream_id",
                                   current.stream_id,
                                   [&](::perfetto::EventContext ctx) {
+                                      if(current.graph_exec_id != 0)
+                                      {
+                                          rocprofiler::sdk::add_perfetto_annotation(
+                                              ctx, "graph_exec_id", current.graph_exec_id);
+                                          rocprofiler::sdk::add_perfetto_annotation(
+                                              ctx, "graph_node_id", current.graph_node_id);
+                                      }
+
                                       for(auto& [counter_id, counter_value] : counter_id_value)
                                       {
                                           rocprofiler::sdk::add_perfetto_annotation(